Remove commented out macros
The macros for Q_Q and Q_D were commented out because the variables
that they create were not used. This is the case because we do not need
the Q_D for the top level class in an object hierarchy. If we were to
subclass both the public and private implementations then we need to
use the macros to get at the proper d_ptr variable and have const
correctness. see http://qt-project.org/wiki/Dpointer for more
information