Fix logic errors when attempting to set a selection in the QComboBox.
The previous logic would allow an item to be inserted into the QComboBox even
if it did not exist in the DataContainer. This logic is more strict in that
if the previous selection or the default from the filter do not exist in
the QComboBox then the QComboBox is set to "no selection". This should
probably be extended to the attribute matrix QComboBox also.