Rewrite class to mimic the actual Vtk source code
The class was rewritten using code lifted from the Vtk project so that the class
should be able to read either CELL_DATA or POINT_DATA in any order. We are
just reading the SCALARS and VECTORS at the moment (both need to be tested). The
code logic uses the same looping style and low leve std::streams classes that
Vtk uses in its implementation.