Compile errors due to spaces in path to the project directory
DOxygen was having issues if there were spaces in the path to the
main DREAM3D project folder. This is because the Doxy file that
got generated was not fully quoting each include path, image path
and the output path. All paths are now quoted properly. Reproduced
on OS X and tested for success there.

updates #249. fixes #249. closes #249