Add derivative structure to geometry interface class
Adding virtual functions necessary for derivative calculations to the geometry
interface class so that all geometries inherit them.  Methods for image, vertex,
edge, and triangle geometries are empty and need to be implemented.  Fixing
incorrect attribute naming for hdf5 arrays holding connectivity information.
Adding check to getPrereqIDataArray to verify if the array can be cast to the
correct type (useful if an array of a specific type is wanted, but the
component dimensions of that array are unknown).

updates #284