Fix incorrect initialization of geometry element sizes
The dimensional size of any tuple of an AttributeMatrix should be initialized to
0 if that size is unknown at the time of dataCheck.  To comply with this
standard, IGeometry objects should also have the size of any of their component
arrays be 0 at initialization in dataCheck, unless the size is known.  This
standard ensures that checks on tuple dimensions of arrays and Geometry sizes
can proceed in preflight for Filters further in the pipeline.