Change FaderWidget functionality to use QPropertyAnimation and QGraphicsEffects
The fade-in and fade-out functionality for linked boolean widgets previously used
a custom implementation to achieve the desired fade effect. The code was switched
to using a QPropertyAnimation hooked to a QGraphicsEffect on the opacity of the
widget. This removes a bunch of custom code for the general case and makes the code
easier to follow. In addition the FilterParameterWidget has been switched to inherit
from QFrame instead of QWidget. This allows us to put a custom CSS Style Sheet border
around the widget. This was not supported for a QWidget. Now when a LinkedBooleanWidget
is triggered the linked FilterParameterWidgdet will fade in with a slightly red
background and then the red disappears after a few seconds.

There is a known issue that after the background goes to a normal color the label
of the FilterParameterWidget is now bold. We are probably inheriting the Stylesheet
from the parent object. This needs to be investigated.