Standardize FilterGroup naming across all internal/external plugins
External plugins now all contain a namespace in their constants file in which
to declare FilterGroups.  Removed unnecessary namespace in Plugin.cpp file in
all plugins (an identical namespace appears in the plugin constants file).
Using spaces in the human labels for all FilterGroups for consistency with
human labels of filters.