Standardizing filter headers
All filter headers now conform to a uniform template.  The template can be found
in Source/Applications/DevHelper/Template/Filter/Filter.h.in, and is the same
template used to generate filters in FilterMaker. The main aspect of the
template is uniform Doxygen for inherited functions, referencing the superclass
AbstractFilter.