Disabling RDF fitting in GenerateEnsembleStatistics
The RDF fitting in GenerateEnsembleStatistics relies on the FindFeatureClustering
filter to generate the necessary data, but that filter is not slated for release
with v6.0. This functionality is being disabled in GenerateEnsembleStatistics
for now.  The filter works correctly otherwise: was able to generate a good
synthetic structure using the statistics from the Small IN100 data set.