Bugfix for FindElementNeighbors in GeometryHelpers
Previously, generating the Triangle Neighbors List in the Generate
Geometry Connectivity Filter resulted in the code silently failing.  No
neighbors were ever recognized.  Additionally, even if the loop were
okay, the storage list was never allocated and would also silently fail.
This fix corrects the minor typo in the for loop in GeometryHelpers.hpp
and also adds another overloaded version of allocateLists to support the
templated type of this list so that allocation can occur.   Not 100%
sure if  the latter fix is necessarily the best course of action, but it
fixes the problem regardless.