Standardize filter defaults
Filter DataContainer default names now have a consistent naming convention that
is [GeometryName]DataContainer. For example, if a filter needs an Image geometry,
the incoming default DataContainer names are "ImageDataContainer". Also changed
for creations of anything to keep it all consistent. This change was made
because with the previous generic name "DataContainer" it was difficult to keep
track of which DataContainers had which types of geometry, and also since the
names were the same the wrong defaults may end up entered in the selection
widgets.