Fix slowness when populating the Error Table.
The issue was that the error table was being built up one row at a time and the QTableWidget
had to keep incrementing the size of the model for each error message. This takes
a 'long' time to accomplish. We now cache all the messages in an internal variable
until a signal is caught that tells the IssuesDock Widget to display the error messages.
The speed up allows filters to be added near real time to the pipeline without
noticeable slow down in the GUI caused by the preflights.