Fix reading for the PhaseTypeSelectionWidget
The PhaseTypeSelectionWidget was using the incorrect property from the filter
as a default, meaning the data array path was never properly set. This has been
addressed, and the logic in the PhaseTypesSelectionWidget now behaves the same
as the AttributeMatrixSelectionWidget. These changes were necessary to get a
working prebuilt pipeline that uses the Generate Ensemble Statistics filter.

updates #368