Add warnings for unsigned types to the FindDifferenceMap filter
The FindDifferenceMap filter simply subtracts the second array from the first
array, which could result in underflow if the second arrays is much larger than
the first and the arrays are of unsigned type. Warnings have been added to the
documentation and to the filter to notify the user of this potential issue.