Fix several issues in the surface mesh sampling filters
The regular grid surface mesh sampling filters assumed an origin of (0,0,0),
which would cause issues when trying to sampling a triangle geometry with a
different origin. The user can now specify the origin of the sampling volume.
Documentation has been updated accordingly. Also, the uncertain grid sampling
routine did not generate a correct volume of points (it instead generated points
localized along a line). This has been corrected, and the filter now produces
the correct set of "uncertain" sampling points.