Fix potential division by zero error in BetaOps class
In the calculateCorrelatedParameters function of the BetaOps class, it was
possible to cause a division by zero if all incoming data as zero, resulting
in NaN return values. Issue has been addressed by checking if the stddev is 0
before computing the alpha/beta parameter values.  Alpha/beta parameters are set
to 0 if the stddev is 0.