Modify 2D shape calculations to avoid infinity/NaN results
The FindShapes filter could possibly generate infinity/NaN results for axis
lengths and aspect ratios.  This usually occurred with very small features
composed of 1 or 2 voxels.  This issue has been addressed by keying of the
feature volume (which originally was not calculated in 2D) to avoid the
division by zero.  A more sophisticated solution would most likely have to
account for potential floating point overflow due to all the successive
additions that occur in the moment calculations.