Fixing JSON parsing issues when reading filter parameters.
DREAM.3D now writes 64 bit integers as strings instead of double values in the
JSON pipeline files. When reading the JSON files additional logic was inserted to
test if the JSON value is a double or string and appropriate conversions are then
used. This allows producers of JSON files to continue to use the native JSON double
object instead of having to update all of their codes to use JSON strings.

Additional logic was inserted to test for the proper ranges of 32 bit floating point
numbers and then correctly parse the JSON value into the floating point variable.

The Unit test was updated to factor out the Integer and Floating point tests into
their own test functions.

updates #432. Fixes #432. closes #432