Add checks for NULL geometries into the various selection widgets
The selection widgets can now filter on the geometry type, but were not checking
to see if the geometry was NULL before asking for the type, which could cause
a crash when dealing with DataContainers that have NULL geometries (such as
those from StatsGenerator).  The widgets will now determine if the geometry is
NULL.  If so, and if the geometry requirements vector has any requirements set,
the data container will be disabled.  Otherwise, the filtering works as
previously implemented.