Documentation will not show when context clicking on filter in a pipeline or clicking on the help button in the FilterInputWidget
Redesigned the help system to create a singleton instance of a DocRequestManager
that the lower level libraries can use and the high level applications can connect
to. This manager class emits the proper signals that the Application level can subscribe
to and decide how to how the documentation for the requested filter.

updates #372