Update QuickSurfaceMesh to allow the user to select any number of cell arrays to transfer to the created surface mesh
The QuickSurfaceMesh filter now allows the user to select any number of cell
arrays to be transferred to the created triangle geometry. The triangles will
gain the values assigned to the cells from which they were created. Currently,
only single-dimensional arrays (i.e., scalars or vectors, but not N x M matrices),
can be transferred. This is due to a limitation in Xdmf hyper slab output in
DREAM.3D; see issue #305 for details.