Fix potential crashing issue due to index out of bounds in MatchCrystallography
Previously, the MatchCrystallography filter was observed to crash randomly and
rarely. The issue was finally caught while running in debug, and seems to be due
to a random number generated as an index. However, based on the code used to
generate the random number, it shouldn't even be possible to get an out-of-bounds
index. Still, fixing the issue always since it definitely happened.  Also note
that there was a check for this, but it was in the wrong place. Leaving the original
check in for more proofing against crashing.