Add a Contributors file to the root repo folder
Adding a file that lists some quick details for contributions made by developers
external to BlueQuartz. This list may be incomplete or contain errors; if so,
developers should submit pull requests with fixes.  If a developer wants to have
their contributions acknowledged in this list, then the necessary changes should
be added with any submitted pull requests.