Add signals to initiate preflights from the StatsGeneratorWidget
The lower level phase and plot widgets now emit when a user modifies a parameter,
which is percolated up by the StatsGeneratorWidget which emits parametersChanged(),
causing a preflight. This in turn gets the values in the widget back down into
the filter without the need to click the update pipeline button. The RDF widget
still needs to be hooked up correctly. Also, need to figure out how to handle
the bulk load options.