Add preflight error checking when selecting bulk load for the ODF
The StatsGenODFWidget will now emit a signal and pass whether it was properly able
to import if the bulk load option is selected. Proper checking is done to ensure
we do not have to re-read the file if it was read correctly at the start. The signal
is caught by a slot in the phase widgets, which uses the parameter passed to set
a local member variable that is checked in the beforePreflight() function at the
StatsGeneratorWidget level, which gets the order right for setting error conditions
in the actual filter.