ENH: Deactivate automatic ITK IO factory registration
By default, ITK IO factory modules are automatically registered
with targets depending on ITK. This is a convenient method but
it does not allow the developer to select the order in which
the different IO modules are registered, and therefore which IO
module is used if multiple modules can read and write a format.
The automatic IO factory registration has been disable to allow
control on the order in which the IO modules are loaded.