Extend the FindSizes filter to operate on any geometry
The FindSizes filter will now operator on feature that belong to any geometry.
Update the documentation to reflect the new functionality. Also add a unit
test to ensure the filter functions properly for all geometries.

This extension requires the additions for SIMPL for IGeometry element sizes.
To break a circular dependence, the API of GeometryMath had to include forward
declares, so places where that API was used (e.g., SampleSurfaceMes) have been
updated to reflect this.