Avizo Exporters do not write correctly formatted files.
Do to the use of QDataStream both the export of binary and ASCII files
are not working correctly. Convert both classes to use FILE* pointers and
raw writing using fprintf and fwrite instead. Add additional parameters
to the header including DREAM.3D version and the Feature Ids Path.

updates #677
fixes #677