Experimental code to find a Triple line in a surface mesh.
This code is here as a reference only and is not actively used. If the code is used
it should be updated to use the actual vertices that are created for the triangle
mesh and share the vertex list between the edge and triangle geometry. This would
cause issues when reading a saved file back into DREAM.3D as there is no way to
link a vertex list as being shared among multiple geometries.