Implement second strategy for parallelizing the sampling of the surface mesh.
If the number of features is less than the number of cores on the hardware then
parallelize over the number of triangle points instead of the number of features.
This allows full use of all the cores on the system.

Also implement some progress feedback for this version of the parallel implementation.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>