Fixed calculation of Volume Fractions in the FindVolFractions filter. (#882)
Another instance of accumulating values in a float that exceed the limits of
floating point data. The solution is to calculate the total number of cells for
each phase then use that value casted to a double as the numerator in the division.
The output values are not correct.

First reported by Helal Chowdhury <helalme111@gmail.com>

Fixes #881. Closes #881.

Fixes #881. Updates #881

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>