Merge Changes to insert atom feature (#899)
* fix uninitialized variable

QVector<BoolArrayType::Pointer> m_InFeature was not defined past the value="start". This caused a seg fault crash when the feature number is high (>200). I'm guessing that the compiler preallocates the array for a few values past "start", which is why it does not crash when the total feature number is low. Fix this bug/type by changing the value to "iter"

* Add cubic crystal basis

Add cubic crystal which is basically adding an atom at the +1/4,+1/4,+1/4 location for each original atom.

* change from double to float

Signed-off-by: MicroClue <whsiao1995@hotmail.com>