Fix issue where Feature Info Reader would not read all data in the file
Comment lines within the data section would cause confusion in the reader
The original algorithm just read the number of lines equal to the number of
features declared in the header. For various reasons this will not work
correctly. The algorithm was changed to read all lines in the file with
the caveat that the feature id value cannot be larger than the max value found
in the Cell AttributeMatrix/FeatureIds array.

updates #930

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>