Moving CaclculateMisorientationInternal into base class
Most classes have the exact same code. Better to update the API to allow for
virtual methods and override where needed.

Fix error in CubicOps that was always returning an Axis Angle with a 0.0 angle

Refactor code in EbsdSegmentFeatures to not create the LaueOpsVector inside a tight
loop. The variable was made an instance variable instead.

All instances of QVector<LaueOps::Pointer> were converted to std::vector<LaueOps::Pointer>

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>