Revert part of last commit to FindAvgOrientations.h
The order of the Python bindings are important as the pythonic interface
arguments are generated in the order of the macros in the header file. By
changing the order of the macros, the order of the arguments are changed
and any code that uses the pythonic interface probably breaks, or has subtle
errors.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>